Clean comments in vrd_trainer, removed shuffle from datalayer (because there's already a shuffle flag in dataloader). Fix smart_frequency

+++ utils.py
@@ -131,7 +131,7 @@ class LeveledAverageMeter(object):
 
 # Smart frequency = a frequency that can be a relative (a precentage) or absolute (integer)
 def smart_fequency_check(iter, num_iters, smart_frequency):
-  if int(smart_frequency) == 0: return False
+  if float(smart_frequency) == 0.0: return False
   if isinstance(smart_frequency, int):
     abs_freq = smart_frequency
   else:
